BT timer code fixed.

+++ MicroFrameworkPK_v4_3/DeviceCode/Targets/Native/SmartFusion2/DeviceCode/drivers/bluetooth/btcore/btstack_run_loop.h
@@ -85,7 +85,7 @@ typedef struct btstack_data_source {
 typedef struct btstack_timer_source {
     btstack_linked_item_t item; 
     // timeout in system ticks (HAVE_EMBEDDED_TICK) or milliseconds (HAVE_EMBEDDED_TIME_MS)
-    uint32_t timeout;
+    uint64_t timeout;
     // will be called when timer fired
     void  (*process)(struct btstack_timer_source *ts); 
     void * context;
